Rename namespace [skip ci]
